replace ui classes to separate package .ui Add description comment
