+++ app/src/main/java/com/shevart/androidcorelearn/different_test_tasks/test_task_1/ShortDescriptionOfTheTestTask.java
@@ -1,11 +1,15 @@
 package com.shevart.androidcorelearn.different_test_tasks.test_task_1;
 
+import com.shevart.androidcorelearn.different_test_tasks.test_task_1.ui.TestTask1Activity;
+
 /**
  * I got test task and there is solution for this task in
  * <b>com.shevart.androidcorelearn.different_test_tasks.test_task_1</b> package.
  *
  * This is link to the test task -
  * https://docs.google.com/document/d/1G7jLw3mRvHZeypZfNgwWjrd-HuNZp0AhLKETs-GbqQI/edit?usp=sharing
+ *
+ * Start point for explore this sample code is {@link TestTask1Activity}
  */
 public final class ShortDescriptionOfTheTestTask {
     private ShortDescriptionOfTheTestTask() {
