+++ app/src/main/java/com/shevart/androidcorelearn/different_test_tasks/test_task_1/ui/PrimeNumbersRVAdapter.java
@@ -1,4 +1,4 @@
-package com.shevart.androidcorelearn.different_test_tasks.test_task_1;
+package com.shevart.androidcorelearn.different_test_tasks.test_task_1.ui;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
