+++ app/src/main/res/layout/activity_test_task1.xml
@@ -4,20 +4,11 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".different_test_tasks.test_task_1.TestTask1Activity">
+    android:background="@android:color/black"
+    tools:context=".different_test_tasks.test_task_1.ui.TestTask1Activity">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/rvPrimeNumbers"
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1" />
-
-    <!--<Button-->
-        <!--android:id="@+id/btStart"-->
-        <!--android:layout_width="match_parent"-->
-        <!--android:layout_height="wrap_content"-->
-        <!--android:paddingLeft="16dp"-->
-        <!--android:paddingRight="16dp"-->
-        <!--android:layout_gravity="center"-->
-        <!--android:text="@string/start" />-->
+        android:layout_height="match_parent" />
 </LinearLayout>
